Moved some files into their own folders to make directory a little more organised
