+++ src/App.jsx
@@ -2,9 +2,9 @@ import { Route, Routes } from "react-router-dom"
 import Homepage from "./components/Homepage"
 import UserPage from "./components/users/UserPage"
 import AlbumPage from "./components/albums/AlbumPage"
-import GetEmailForSignUpPage from "./components/users/sign-in/GetEmailForSignUpPage"
+import GetEmailForSignUpPage from "./components/users/sign-in/email-sign-up/GetEmailForSignUpPage"
 import SignInPage from "./components/users/sign-in/SignInPage"
-import CompleteSignUpPage from "./components/users/sign-in/CompleteSignUpPage"
+import CompleteSignUpPage from "./components/users/sign-in/email-sign-up/CompleteSignUpPage"
 import UserSettingsPage from "./components/users/settings/UserSettingsPage"
 import LoadingPage from "./components/LoadingPage"
 import Header from "./components/header/Header"
@@ -23,7 +23,7 @@ function App() {
       <Route path="/search" element={<SongsList/>}/>
       <Route path="/albums/:album_id" element={<AlbumPage/>}/>
       <Route path="/sign_in" element={<SignInPage/>}/>
-      <Route path="/create_account" element={<GetEmailForSignUpPage/>}/>
+      <Route path="/sign_in/create_account" element={<GetEmailForSignUpPage/>}/>
       <Route path="/complete_signup" element={<CompleteSignUpPage/>}/>
       <Route path="/users/settings" element={<UserSettingsPage/>}/>
       <Route path="/loading" element={<LoadingPage/>}/>
