+++ src/components/users/sign-in/SignInPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import StyledLink from "../../styling/StyledLink"
-import EmailAndPasswordSignIn from "./EmailAndPasswordSignIn"
-import GoogleSignIn from "./GoogleSignIn"
+import EmailAndPasswordSignIn from "./email-sign-up/EmailAndPasswordSignIn"
+import GoogleSignIn from "./google-sign-in/GoogleSignIn"
 import Loading from "../../Loading";
 
 function SignInPage(){
@@ -26,7 +26,8 @@ function SignInPage(){
         <br/>
         <GoogleSignIn setIsLoading={setIsLoading} setError={setError}/>
         {error === "Google sign in error" ? <p>Error signing in with Google. Please try again later.</p> : null}
-        <p>New to this site? <StyledLink to="/create_account">Create an account</StyledLink></p>
+        <p>Forgot password? <StyledLink to="/sign_in/reset_password">Reset password</StyledLink></p>
+        <p>New to this site? <StyledLink to="/sign_in/create_account">Create an account</StyledLink></p>
     </section>)
 }
